requirements.txt
streamlit
langchain
python-dotenv
langchain_community
faiss-cpu
transformers
tavily-python
sentence_transformers
openai
langchain_openai
torch
tiktoken